MscmpSystMcpPerms.Types (mscmp_syst_mcp_perms v0.1.0)

Types used by the MCP Perms component.

Types

access_account_perm_role_params()

@type access_account_perm_role_params() :: %{
  access_account_id: MscmpSystAuthn.Types.access_account_id(),
  perm_role_id: MscmpSystPerms.Types.perm_role_id()
}

A map defining the data elements required for creating Access Account Permission Role Assigns records.

Attributes

access_account_id - the record ID of the Access Account record being granted a Permission Role.

perm_role_id - the record ID of the Permission Role which is to be granted to the identified Access Account.